Move folder, fix Arduino and EEPROM includes
