+++ Midi_In_YM2151/Common.h
