+++ Midi_In_YM2151/EPROMManager.cpp
@@ -16,7 +16,7 @@
 *	along with this program.If not, see <http://www.gnu.org/licenses/>.
 */
 
-#include  <EEPROM/src/EEPROM.h>
+#include  <EEPROM.h>
 
 #include "EPROMManager.h"
