+++ Midi_In_YM2151/NotePool.cpp
