+++ Midi_In_YM2151/YM2151Driver.h
@@ -20,7 +20,7 @@
 #define _YM2151DRIVER_h
 
 #if defined(ARDUINO) && ARDUINO >= 100
-	#include "arduino.h"
+	#include <Arduino.h>
 #else
 	#include "WProgram.h"
 #endif
